Move dashboard store related modules to web/store/dashboard
All redux store related code will be placed under web/store now.
