+++ ng/src/web/pages/cpes/dashboard/loaders.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Loader, {
   loadFunc,
   loaderPropTypes,
-} from '../../../components/dashboard2/data/loader';
+} from 'web/store/dashboard/data/loader';
 
 export const CPES_CREATED = 'cpes-created';
 export const CPES_SEVERITY = 'cpes-severity';
