+++ ng/src/web/pages/cves/dashboard/loaders.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Loader, {
   loadFunc,
   loaderPropTypes,
-} from '../../../components/dashboard2/data/loader';
+} from 'web/store/dashboard/data/loader';
 
 export const CVES_SEVERITY = 'cves-severity';
 export const CVES_CREATED = 'cves-created';
