+++ ng/src/web/pages/os/dashboard/loaders.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Loader, {
   loadFunc,
   loaderPropTypes,
-} from '../../../components/dashboard2/data/loader';
+} from 'web/store/dashboard/data/loader';
 
 export const OSS_SEVERITY = 'oss-severity';
 export const OSS_VULN_SCORE = 'oss-most-vulnerable';
