+++ ng/src/web/pages/ovaldefs/dashboard/loaders.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Loader, {
   loadFunc,
   loaderPropTypes,
-} from '../../../components/dashboard2/data/loader';
+} from 'web/store/dashboard/data/loader';
 
 export const OVALDEFS_CLASS = 'ovaldefs-class';
 export const OVALDEFS_CREATED = 'ovaldefs-created';
