+++ ng/src/web/pages/results/dashboard/loaders.js
@@ -25,7 +25,7 @@ import React from 'react';
 import Loader, {
   loadFunc,
   loaderPropTypes,
-} from '../../../components/dashboard2/data/loader';
+} from 'web/store/dashboard/data/loader';
 
 export const RESULTS_DESCRIPTION_WORDCOUNT = 'results-description-wordcount';
 export const RESULTS_SEVERITY = 'results-severity';
