+++ ng/src/web/store/dashboard/settings/__tests__/reducers.js
